broke the note scriptable out and note maker editor now adds to the notebook
