+++ Assets/Editor/NotebookMakerEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af9087ace8a4602439495d0256e5f19a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
