+++ Assets/Resources/NewNotebook.asset
@@ -8,13 +8,28 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 0b60dfdbfac478a45857029bedfdbfd9, type: 3}
   m_Name: NewNotebook
-  m_EditorClassIdentifier: Assembly-CSharp:Zach:NotebookScriptable
+  m_EditorClassIdentifier: 
   notes:
-  - {fileID: 114886333654802664}
-  - {fileID: 114326449441629776}
---- !u!114 &114326449441629776
+  - {fileID: 114868890011481554}
+  - {fileID: 114935185357318310}
+  - {fileID: 114846225819220846}
+--- !u!114 &114846225819220846
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 15b2241fbe291a3408bed7424e2f12c2, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  isEnabled: 0
+  noteName: kiwami
+  data: it means extreme
+--- !u!114 &114868890011481554
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
@@ -28,7 +43,7 @@ MonoBehaviour:
   isEnabled: 0
   noteName: 
   data: 
---- !u!114 &114886333654802664
+--- !u!114 &114935185357318310
 MonoBehaviour:
   m_ObjectHideFlags: 0
   m_CorrespondingSourceObject: {fileID: 0}
